requirements.txt
# Core LLM and Agent
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.1
openai>=1.0.0

# Web UI
chainlit>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
plotly>=5.0.0

# Utilities
python-dotenv>=1.0.0
tabulate>=0.9.0
pydantic>=2.0.0
